Updated fieldtype from Data to Text for 'description' fieldname
In the database, the field size is defined as Text for Description but in the java script it is being used as Data field. Hence the user are not able to enter the full text when assigning a ToDo to another use. Please accept this change as it will be very helpful for the user. Thanks.
Mayur