Allow SSL connections to a remote database server
Add the following options in site_config.json to use ssl
- db_ssl_ca = "/path/to/ca/cert.pem"
- db_ssl_cert = "/path/to/ssl/cert.pem"
- db_ssl_key = "/path/to/ssl/key.pem"
The files mentioned should be stored on the App server in the directory mentioned.

This commit allows for basic ssl connections. X509 is still not supported.

If the above options are not set, the connection will not use an encrypted connection, and connect normally.