Combine Login with LDAP with normal login
There was a separate Login with LDAP button when LDAP integration is enabled which will confuse normal user.
This enhancement removed Login with LDAP button and related code. Add lines of code letting user can login with LDAP when failed authentication with local database.
Which means when a user perform a login, LoginManager will authenticate the user with local database first, then attempt using LDAP if local authentication failed and LDAP is enabled.