Check if subject is default when sending mail
When sending an e mail by a standard reply / email template, using the Communication Composer, the code checks if subject is empty. It should check as well, if the subject is in the default format, because otherwise the subject of the Standard Reply / E Mail Template will never be used.