fix: identify production mode for PDF generation
In regards to PDF generation, if developer_mode was set to true, the PDF
generated used to lack css styling. This was because the URL to the CSS
assets used to contain the webserver port number, but given the system
was running in production, the URL was invalid, thereby not being able
to load CSS assets. This fix does not rely on the value of
developer_mode to identify if the system is running in production,
rather, it uses the value of restart_superviosor or systemd_on_update,
to check for the same.