Set visibility of body by class rather
When including an external JS code, for example, the popular Rocket Chat Live Chat Widget, the external application (in this case Rocket Chat Live Chat) resets the inline styles of the body tag and makes the webform disappear quickly after it appeared.

As this is a strange behavior of the Rocket Chat Live Chat Widget, I will report it there. However, in this case, in my opinion it's still more elegant to set the visibility of the body tag by the bootstrap class 'show'.