fix: disable smart shrinking
Adding this option was a fix for a user who manually changed the version of wkpdftohtml.
With this addition all standard installs (wkhtmltopdf 0.12.3 (with patched qt)) project 20% more surface
on which background images for pdf invoices no longer work. we cannot send invoices anymore and this
is not fixable by pdf (html) options because the 'zoom' function is not implemented

please remove and ask original commitor to change the addition to a setting by print settings.