Add support for sending JSON to SMS Settings
Many SMS gateways expect data in raw JSON and do not accept plain
form data. This change enables this by detecting a "Content-Type"
header in the parameters with a value of "application/json", and
in this case decoding the message text to unicode as expected by
the `requests` module before passing the data to the `json`
parameter.