fix: check if user exists in the system before commenting
comments on blog posts allows users to comment with random email
addresses. if the user with the email exists in the system, it creates
the comment as that user. so to prevent forging comments under some
other user, we add a check to see whether the user exists and is
currently logged in.

Signed-off-by: Chinmay D. Pai <chinmaydpai@gmail.com>