fix: Bypass ORM for removing apps from installed_apps list
Scenario:
a site (on older version) has a custom app xyz installed.
this site is to be restored on a new bench (without xyz) and migrated to recent version.

Problem:
`bench migrate` fails because xyz is not available.
`bench remove-from-installed-apps` might fail, because code is not compatible with the database.