fix: wrong type when writing with csv writer
Unlike the get_untranslated function, in which the file is opened and
written with the `write()` function, here the `csv.writer()` is used,
causing errors when writing encoded strings. Therefore we remove the
encoding step (The csv writer will take care of it) and we also not open
the file with binary mode.

(cherry picked from commit ec5851abf7831899ef3986e453c323b42c3b6a1b)