fix(backups): bench backup command should raise an error when a backup fails
Using frappe.commands.popen instead of frappe.utils.execute_in_shell, because the latter doesn't raise any error in case of a non-zero exit code.

The exit code of the pipeline is the exit code of the last command (gzip), So, if mysqldump fails for any reason it isn't registered as a failure.
Using bash -o pipefail as a fix for this (/bin/sh doesn't support this).