fix: set default gender to 'Non-Confirming' if no gender information is provided by OAuth provider. (temporarily fix)
Note: This fix is mainly for a temporarily fix on a error blocking OAuth flow being finished.

There are more problem in file `frappe/frappe/utils/oauth.py:237`, which is a misuse of `frappe.new_doc(). Second parameter is being filled with a dict object, it is for parent doctype, which should be a doctype object rather than a dict.

Maybe something wrong happened in git workflow. Please check commit a2fba771

This fix is mainly a temporarily approach, but also reasonable on long term (providing a reasonable default value).