Update webhook.py
I am unable to test this because I don't have access to a non production machine.

The json validation won't allow useful, and arguably vital jinja script.

This use case for being able to use jinja might be the ability to parse through all elements of a table within a doc. 

Example:  Items Doctype (From ERPNext)

{
    "name":"{{doc.name}}",
    "uoms":[
        {
            {% for uom in doc.uoms %}
            "name":"{{uom.name}}"
            {% endfor %}}"
        }
    ]
}

This way we can select names exclusively. There are other functionalities that validation was blocking as well.