feat: @frappe.write_only
When would you want to use it? Typically if you want to be sure that
"this" particular thing will be make WRITES to the database. If this
were to be called via a function wrapped with @frappe.read_only. This
will momentarily switch the connection from replica to primary

(cherry picked from commit 61d067710ffd1819f0ed26548ea64d915a8cbc45)