fix: Erroreneous query building when column exists in matching condition and set field value
Given how we're still trying to refactor set_value for v14; instead of
diverging the fixes for each major version I'd rather change v12 to be
more similar to the next two versions instead.

So, I've pasted the changes from v14 as is. Repurcussions of this
include an extra select query on every set_value call...just as it is in
v13 and v14.

Once frappe.db.set_value is refactored, I plan to backport the perf
improvements to v13 and v12 so that the extra query isn't required.

(cherry picked from commit ed37d5299583096ade2d685f6ae1feefc0ca8408)