build: Bump python_requires to match documentation
python_requires wasn't changed from a while due to an oversight
but Frappe v13 has always been >= PY37
refs:
* https://frappeframework.com/docs/v13/user/en/installation#pre-requisites
* https://github.com/frappe/frappe/blob/version-13/.github/workflows/patch-mariadb-tests.yml#L27

This was highlighted when we last bumped iPython to fix a security
vulnerability (ref:
https://github.com/frappe/frappe/commit/740db83bd122bd920eb4e285d8525ebdd165cb21#diff-4d7c51b1efe9043e44439a949dfd92e5827321b34082903477fd04876edb7552R33)
when PY36 installs failed. The workaround for this wouldve been to
simply ignore updating this particular dependency while running through
pip, but Frappe doesn't have to support it.