fix(Kanban Board): Clear user settings & dpctype cache on_change
Deletion of kanban boards may lead to 404s on clicking on the Kanban
option in the views. This is due to the cache not being synced with the
DB changes. Changed the hook to on_change since it gets triggered even
in db_set.

(cherry picked from commit fd0564953b0d5dd32d563617b1894e48d0aa8c55)