fix: preserve list style when cleaning HTML
Currently, orderered (OL) and unordered lists (UL) are listed as allowed
tags for `clean_html`. But the editor (Quill) does not generate UL
elements for unordered lists but generates an OL element with
`data-list="bullet"` instead. The list style is lost during cleaning.

To preserve the list style, `data-list` has been added to the list of
allowed attributes.

Refs: #14104, #14171, quilljs/quill#3014