refactor: remove dead flag db.read_only
This was added in last DB refactor but it does nothing, it was probably
supposed to do something with the connection pool but to best of my
knowledge "read only" is not a property of a connection.

It can be achieved with users who only have read access, that however
isn't implemented anywhere.

Removing this for now.

(cherry picked from commit ea7fbb2c10752db929fc038069884f92e0172533)

# Conflicts:
#	frappe/database/__init__.py