fix: correct index re-syncing
The implementation of syncing unique and non-unique index depended on
index names which used to be different before because of that there's
tendency to incorrectly identify index.

This PR adds a separate util for checking if a column has index without
relying on naming convention. It just goes and checks if there's any
index with that column in it, hence far more reliable.

(cherry picked from commit cbe4b591000599e92f14f6451e0a88d28d5b89ab)

# Conflicts:
#	frappe/database/mariadb/database.py
#	frappe/database/mariadb/schema.py