chore: give cli invocations of dbtools a dedicated interface
This PR removes technical debt. It does not add any features in itself.

By capturing the multiple ways in which python code interfaces with the
database CLI tooling and consolidating the command construction in a
single place / function, this PR increases confidence of for future
refactorings and interface amendments

While this commit does not make such amendments, a follow up PR may,
for example, add full unix socket support to the db connections.