(fix): initial sync based in initial_sync_count
In the case of initial_sync/ reindexing, we have unexpected behavior, as we were fetching too many messages, as email_list (holding uids) is sorted asc, where the smallest uid = the oldest email. Therefore, when syncing ALL emails, we (a) have to reverse the order and (b) check for the initial_sync_count.