chore: move db defaulting logic out of the low level db interface
This commit removes technical debt. The fact that the name of a database
equals the user is a high-level frappe framework concern.

By lifting this default logic to the appropriate (high-level) interface
the, the code is normalized and future refactorings can be made with
higher confidence.

One such future refactoring involves implementation of unix domain
socket support which simplifies user / password management. With UDS,
user and password become mostly obsolete which is why the cur_db_name
is being promoted to the only source of truth.

In an attempt to avoid multivariant regression, this refactoring is done
in a separate and preliminary PR. This should also help with review.