fix: temporarily increase recursion depth for safe_eval
This is required because people write long af expression for taxes and
the node transformer roughly visits each node 4ish times. So if you have
a nested if cond that's ~1000 conditions long then it will hit python's
recursion depth limit.

This is a temporary workaround, I'll figure out an alternate fix for it.