fix: workspace don't strip filter info on save.
currently on save of filter ( get_filter_as_json ) is used
to convert it to object with following structure:

```
{
   priority: (2) ['=', 'Medium'],
   status: (2) ['=', 'Open']
}
```

Doing this we lose information of parent doctype.
Missing parent doctype causes field not permitted error.

To fix this, I saved filter as it is and converted it to object when required.