fix: Don't mutate `df` object in frappe.format
When customizing the **Communication** DocType by setting the property `in_list_view` for the field `_user_tags` to `1`, the *fieldtype* of the `_user_tags` docfield can be mutated to `"Tag"` by the `frappe.format` function called during the list view rendering. Then, when opening a Communication form view, the layout tries to render the `_user_tags` field which has the invalid fieldtype `"Tag"`, which crashes the rendering of the form.