fix: remove quotes from IMAP Folder Name (#23242)
Quotes were added around folder name as spaces would cause issue with IMAP.

This extra quotes are also passed to query which sets the
uidnext and uidvalidity. where condition is never met

This values are used on subsequent pulls and without this values,
it is considered first pull and all UNSEEN mails are synced.

So even If user selected All in sync option it will always work as UNSEEN
option.

When UNSEEN is used it will mark all emails as seen on IMAP server when pulled.

(cherry picked from commit 0b4f06c3250be574512d6853906435dc51219618)