fix: set correct recipient when reply to own email (#24256)
* fix: set correct recipient when reply to own email

When Sender clicks on reply to own email, the recipient is set to the Sender's email address instead of the original ( last_email ) recipient.
Added a check if sender is composing a reply to own email and set the recipient to the original ( last_email ) recipient.

* Update communication.js

Co-authored-by: Ankush Menat <ankushmenat@gmail.com>

---------

Co-authored-by: Ankush Menat <ankushmenat@gmail.com>
(cherry picked from commit 141abf9faf93e01cd53e6ce53140f110d2efe99b)