fix: parse error messages in data import
When error is thrown during data import.
The error message is saved in data import log as json string.
Even though we are parsing log.messages as json.
we are not parsing indivisual error messages which are also saved as json string.

added a check if individual error message is string then try to parse it.