feat(web): Add dynamic web template fields
When checking the "Dynamic template" option in the field of a Web Template,
it's content will be rendered before rendering the whole web block.

For example, checking the option for the Markdown web template would allow
users of this block (in a Web Page) to use Jinja templates inside the text.