fix: restoring site breaks when checking backup version (#26186)
* fix: restoring site breaks when checking backup version

The `backup_version` variable exists if and only if the `if match` branch executes, but it is used in a condition in a scope where it may not exist. Changing the variable to `match` leads to correct behaviour.

* fix: restoring site breaks when checking backup version

Co-authored-by: Akhil Narang <me@akhilnarang.dev>

* fix: version check when restoring site should be done from `match` scope

* chore: add a `return None` for better readability

No functional difference, just easier to understand/read the code.

Signed-off-by: Akhil Narang <me@akhilnarang.dev>

---------

Signed-off-by: Akhil Narang <me@akhilnarang.dev>
Co-authored-by: Akhil Narang <me@akhilnarang.dev>
(cherry picked from commit 05766680f7ff8d00b938dad8917ca8617abec35c)