fix: Redirect to Login Page if not signed In
Redirect to Login page if User is not logged in already. Upon successful login, redirect to the page they intended to see.

This is a common pattern for opening tabs to previous sessions where Users get logged out and are shown the "Log in to access this page." message instead.