Fix: Map icon color in dark mode
When toggling to dark mode, all the icons in the dropdown button change to white in the list view, except for the map icon, which remains black.

Upon reviewing the code for this icon, I found that it had a fixed white color assigned (stroke=#111). I replaced it with the variable color value var(--icon-stroke) to ensure it changes appropriately with the theme toggle.

(cherry picked from commit 0ebe4a2448b7ac8d2c083cd5a76e8bb9fa7bebf8)