Update schema.py
When you run `bench --site <site-name> migrate` the default value given for Date data type `Today` is not handled in `schema.py`. That is causing `migrate` to create an error and stop the migration. This commit handles the data type `Date` and default value `Today` - which is the case for example in DocType `Stock Entry` and field `posting_date`