fix: controller path for exporting types
When a DocType's controller is overridden by another app, we tried to write type annotations to the path of the overriding controller. Thus, the type annotations never got exported to the original controller. Now we always use the path of the original controller.

E.g. if ERPNext and HRMS are installed and we edit the Timesheet doctype, the `controller_path` used to be set t o `hrms/overrides/employee_timesheet.py`. Now it is set to `erpnext/projects/doctype/timesheet/timesheet.py`.