perf: Use ContextVars w/ LocalProxy instead of magical namepsaces
Magical namespaces avoid managing contextvars but they do so at cost of
copying that locals dictionary every time it is accessed.

This fix is for demonstration purposes, I'll be attempting a rigourous
fix before merge that covers all locals and provides a namespace for
developers that is as efficient as `_dict` but still request local.